+++ pybcpy/__init__.py
@@ -1,2 +1,2 @@
-
-
+VERSION = "v0.0.17"
+VERSION_add = "this software is in alpha state. refer homepage on github for more"